EUI-4873 reapplied dynamic list values for mid event callbacks
EUI-4873 refactoring


EUI-4873 final fix


EUI-4873 clean up
