+++ src/shared/components/case-editor/services/cases.service.spec.ts
@@ -539,78 +539,6 @@ describe('CasesService', () => {
     });
   });
 
-  describe('setDynamicListDefinition()', () => {
-
-    it('should set data for dynamic lists', () => {
-
-      const callbackResponse = {
-            field_type: {
-              complex_fields: [
-                {
-                  field_type: {
-                    type: 'DynamicList'
-                  },
-                  id: 'complex_dl',
-                  value: {},
-                  formatted_value: {}
-                }
-              ],
-              type: 'Complex'
-            },
-            value: {
-              complex_dl: {
-                list_items: [
-                  {code: '1', value: '1'},
-                  {code: '2', value: '2'}
-                ],
-                value: {code: '2', value: '2'}
-              }
-            }
-      };
-
-      (casesService as any).setDynamicListDefinition(callbackResponse, callbackResponse.field_type, callbackResponse);
-
-      const expected = {
-            field_type: {
-              complex_fields: [
-                {
-                  field_type: {
-                    type: 'DynamicList'
-                  },
-                  id: 'complex_dl',
-                  value: {
-                    list_items: [
-                      {code: '1', value: '1'},
-                      {code: '2', value: '2'}
-                    ],
-                    value: {code: '2', value: '2'}
-                  },
-                  formatted_value: {
-                    list_items: [
-                      {code: '1', value: '1'},
-                      {code: '2', value: '2'}
-                    ],
-                    value: {code: '2', value: '2'}
-                  }
-                }
-              ],
-              type: 'Complex'
-            },
-            value: {
-              complex_dl: {
-                list_items: [
-                  {code: '1', value: '1'},
-                  {code: '2', value: '2'}
-                ],
-                value: {code: '2', value: '2'}
-              }
-            }
-      };
-
-      expect(callbackResponse).toEqual(expected);
-    });
-  });
-
   describe('createChallengedAccessRequest()', () => {
     beforeEach(() => {
       httpService.post.and.callThrough();
