+++ src/shared/components/case-editor/services/cases.service.ts
@@ -22,7 +22,7 @@ import {
   RoleRequestPayload
 } from '../../../domain';
 import { UserInfo } from '../../../domain/user/user-info.model';
-import { HttpErrorService, HttpService, LoadingService, OrderService, SessionStorageService } from '../../../services';
+import { FieldsUtils, HttpErrorService, HttpService, LoadingService, OrderService, SessionStorageService } from '../../../services';
 import { CaseAccessUtils } from '../case-access-utils';
 import { WizardPage } from '../domain';
 import { WizardPageFieldToCaseFieldMapper } from './wizard-page-field-to-case-field.mapper';
@@ -50,11 +50,6 @@ export class CasesService {
     'application/vnd.uk.gov.hmcts.ccd-data-store-api.create-event.v2+json;charset=UTF-8';
   public static readonly V2_MEDIATYPE_CREATE_CASE =
     'application/vnd.uk.gov.hmcts.ccd-data-store-api.create-case.v2+json;charset=UTF-8';
-
-  // Handling of Dynamic Lists in Complex Types
-  public static readonly SERVER_RESPONSE_FIELD_TYPE_COLLECTION = 'Collection';
-  public static readonly SERVER_RESPONSE_FIELD_TYPE_COMPLEX = 'Complex';
-  public static readonly SERVER_RESPONSE_FIELD_TYPE_DYNAMIC_LIST_TYPE: FieldTypeEnum[] = ['DynamicList', 'DynamicRadioList'];
 
   public static readonly PUI_CASE_MANAGER = 'pui-case-manager';
 
@@ -117,79 +112,6 @@ export class CasesService {
       );
   }
 
-  /**
-   * handleNestedDynamicLists()
-   * Reassigns list_item and value data to DynamicList children
-   * down the tree. Server response returns data only in
-   * the `value` object of parent complex type
-   *
-   * EUI-2530 Dynamic Lists for Elements in a Complex Type
-   *
-   * @param jsonBody - { case_fields: [ CaseField, CaseField ] }
-   */
-  private handleNestedDynamicLists(jsonBody: { case_fields: CaseField[] }): any {
-
-    if (jsonBody.case_fields) {
-      jsonBody.case_fields.forEach(caseField => {
-        if (caseField.field_type) {
-          this.setDynamicListDefinition(caseField, caseField.field_type, caseField);
-        }
-      });
-    }
-
-    return jsonBody;
-  }
-
-  private setDynamicListDefinition(caseField: CaseField, caseFieldType: FieldType, rootCaseField: CaseField) {
-    if (caseFieldType.type === CasesService.SERVER_RESPONSE_FIELD_TYPE_COMPLEX) {
-
-      caseFieldType.complex_fields.forEach(field => {
-        try {
-          const isDynamicField = CasesService.SERVER_RESPONSE_FIELD_TYPE_DYNAMIC_LIST_TYPE.indexOf(field.field_type.type) !== -1;
-
-          if (isDynamicField) {
-            const dynamicListValue = this.getDynamicListValue(rootCaseField.value, field.id);
-            if (dynamicListValue) {
-              const list_items = dynamicListValue.list_items;
-              const value = dynamicListValue.value;
-              field.value = {
-                list_items: list_items,
-                value: value ? value : undefined
-              };
-              field.formatted_value = {
-                ...field.formatted_value,
-                ...field.value
-              };
-            }
-          } else {
-            this.setDynamicListDefinition(field, field.field_type, rootCaseField);
-          }
-        } catch (error) {
-          console.log(error);
-        }
-      });
-    } else if (caseFieldType.type === CasesService.SERVER_RESPONSE_FIELD_TYPE_COLLECTION) {
-      if (caseFieldType.collection_field_type) {
-        this.setDynamicListDefinition(caseField, caseFieldType.collection_field_type, rootCaseField);
-      }
-    }
-  }
-
-  private getDynamicListValue(jsonBlock: any, key: string) {
-
-    if (jsonBlock[key]) {
-      return jsonBlock[key];
-    } else  {
-      for (const elementKey in jsonBlock) {
-        if (typeof jsonBlock === 'object' && jsonBlock.hasOwnProperty(elementKey)) {
-          return this.getDynamicListValue(jsonBlock[elementKey], key);
-        }
-      }
-    }
-
-    return null;
-  }
-
   getEventTrigger(caseTypeId: string,
                   eventTriggerId: string,
                   caseId?: string,
@@ -214,7 +136,7 @@ export class CasesService {
       .get(url, {headers, observe: 'body'})
       .pipe(
         map(body => {
-          return this.handleNestedDynamicLists(body);
+          return FieldsUtils.handleNestedDynamicLists(body);
         }),
         catchError(error => {
           this.errorService.setError(error);
